+++ src/main/java/com/kakaotech/team14backend/outer/member/controller/MemberController.java
@@ -3,7 +3,7 @@ package com.kakaotech.team14backend.outer.member.controller;
 import com.kakaotech.team14backend.auth.PrincipalDetails;
 import com.kakaotech.team14backend.common.ApiResponse;
 import com.kakaotech.team14backend.common.ApiResponseGenerator;
-import com.kakaotech.team14backend.inner.member.service.FindMemberInfoService;
+import com.kakaotech.team14backend.inner.member.usecase.FindMemberInfoUsecase;
 import com.kakaotech.team14backend.outer.member.dto.GetMemberInfoResponseDTO;
 import com.kakaotech.team14backend.outer.member.service.MemberService;
 import io.swagger.annotations.ApiOperation;
@@ -21,14 +21,14 @@ import org.springframework.web.bind.annotation.RestController;
 public class MemberController {
 
   private final MemberService memberService;
-  private final FindMemberInfoService findMemberInfoService;
+  private final FindMemberInfoUsecase findMemberInfoUscase;
 
   @ApiOperation(value = "마이페이지 계정 상세 조회")
   @GetMapping("/user/info")
   public ApiResponse<ApiResponse.CustomBody<GetMemberInfoResponseDTO>> getMyPageInfo(
       @AuthenticationPrincipal PrincipalDetails principalDetails) {
     Long memberId = principalDetails.getMember().getMemberId();
-    GetMemberInfoResponseDTO myPageInfo = findMemberInfoService.getMyPageInfo(memberId);
+    GetMemberInfoResponseDTO myPageInfo = findMemberInfoUscase.getMyPageInfo(memberId);
     return ApiResponseGenerator.success(myPageInfo, HttpStatus.OK);
   }
