+++ src/main/java/com/kakaotech/team14backend/outer/point/service/PointService.java
@@ -1,7 +1,7 @@
 package com.kakaotech.team14backend.outer.point.service;
 
 import com.kakaotech.team14backend.inner.member.model.Member;
-import com.kakaotech.team14backend.inner.member.usecase.FindMemberUsecase;
+import com.kakaotech.team14backend.inner.member.service.FindMemberService;
 import com.kakaotech.team14backend.inner.point.model.UsePointDecider;
 import com.kakaotech.team14backend.inner.point.usecase.UsePointUsecase;
 import com.kakaotech.team14backend.inner.point.usecase.ValidatePointByPopularPostUsecase;
@@ -17,13 +17,14 @@ public class PointService {
 
   private final ValidatePointByPopularPostUsecase validatePointByPopularPostUsecase;
   private final UsePointUsecase usePointUsecase;
-  private final FindMemberUsecase findMemberUsecase;
+  private final FindMemberService findMemberService;
 
-  public String usePointByPopularPost(UsePointByPopularPostRequestDTO usePointByPopularPostRequestDTO, Long memberId) {
+  public String usePointByPopularPost(
+      UsePointByPopularPostRequestDTO usePointByPopularPostRequestDTO, Long memberId) {
     validatePointByPopularPostUsecase.execute(usePointByPopularPostRequestDTO);
-    Member member = findMemberUsecase.execute(memberId);
+    Member member = findMemberService.execute(memberId);
     Long point = UsePointDecider.decidePoint(usePointByPopularPostRequestDTO.postLevel());
-    usePointUsecase.execute(member,point);
+    usePointUsecase.execute(member, point);
     return member.getInstaId();
   }
